Add caching for DB table structure operations
Similar to ac6f81d3 for table exists checks.

By remembering the table structure, we don't need to look it up each time we want to insert or update a row. Saves a little bit of time off mass-inserts/updates.

Until we can improve `alterTable()`, the cached table structure is cleared whenever a change is made to the table.

Added:
- Method `setTableStructure()` to store copy of table schema on PDO instance
- Method `queryTableStructure()` to prepare and run actual table schema retrieval

Changed:
- Method `tableStructure()` to delegate and remember lookup
- Method `alterTable() to clear cached table structure